Restructued build system to support a larger project 	modified:   .gitignore 	modified:   Makefile 	new file:   config.mk 	new file:   fdisk.conf 	new file:   hdd/home/test.txt 	new file:   kernel/Makefile 	renamed:    src/boot/boot.asm -> kernel/boot.asm 	renamed:    src/kernel/cpu/gdt.c -> kernel/cpu/gdt.c 	renamed:    src/kernel/cpu/idt.c -> kernel/cpu/idt.c 	renamed:    src/kernel/cpu/interrupt.asm -> kernel/cpu/interrupt.asm 	renamed:    src/kernel/cpu/irq.c -> kernel/cpu/irq.c 	renamed:    src/kernel/cpu/isr.c -> kernel/cpu/isr.c 	renamed:    src/kernel/drivers/ata.c -> kernel/drivers/ata.c 	renamed:    src/kernel/drivers/dma.c -> kernel/drivers/dma.c 	renamed:    src/kernel/drivers/fdc.c -> kernel/drivers/fdc.c 	renamed:    src/kernel/drivers/io_port.c -> kernel/drivers/io_port.c 	renamed:    src/kernel/drivers/keyboard.c -> kernel/drivers/keyboard.c 	renamed:    src/kernel/drivers/pci.c -> kernel/drivers/pci.c 	renamed:    src/kernel/drivers/rtc.c -> kernel/drivers/rtc.c 	renamed:    src/kernel/drivers/timer.c -> kernel/drivers/timer.c 	renamed:    src/kernel/drivers/vga.c -> kernel/drivers/vga.c 	renamed:    src/kernel/fs/devfs.c -> kernel/fs/devfs.c 	renamed:    src/kernel/fs/devices/null.c -> kernel/fs/devices/null.c 	renamed:    src/kernel/fs/devices/zero.c -> kernel/fs/devices/zero.c 	renamed:    src/kernel/fs/msdospart.c -> kernel/fs/msdospart.c 	renamed:    src/kernel/fs/vfs.c -> kernel/fs/vfs.c 	renamed:    src/kernel/include/cpu/gdt.h -> kernel/include/cpu/gdt.h 	renamed:    src/kernel/include/cpu/idt.h -> kernel/include/cpu/idt.h 	renamed:    src/kernel/include/cpu/irq.h -> kernel/include/cpu/irq.h 	renamed:    src/kernel/include/cpu/isr.h -> kernel/include/cpu/isr.h 	renamed:    src/kernel/include/display.h -> kernel/include/display.h 	renamed:    src/kernel/include/drivers/ata.h -> kernel/include/drivers/ata.h 	renamed:    src/kernel/include/drivers/dma.h -> kernel/include/drivers/dma.h 	renamed:    src/kernel/include/drivers/fdc.h -> kernel/include/drivers/fdc.h 	renamed:    src/kernel/include/drivers/io_port.h -> kernel/include/drivers/io_port.h 	renamed:    src/kernel/include/drivers/keyboard.h -> kernel/include/drivers/keyboard.h 	renamed:    src/kernel/include/drivers/pci.h -> kernel/include/drivers/pci.h 	renamed:    src/kernel/include/drivers/rtc.h -> kernel/include/drivers/rtc.h 	renamed:    src/kernel/include/drivers/timer.h -> kernel/include/drivers/timer.h 	renamed:    src/kernel/include/drivers/vga.h -> kernel/include/drivers/vga.h 	renamed:    src/libc/include/hashmap.h -> kernel/include/dsa/hashmap.h 	renamed:    src/libc/include/list.h -> kernel/include/dsa/list.h 	renamed:    src/libc/include/tree.h -> kernel/include/dsa/tree.h 	renamed:    src/kernel/include/fs/fs.h -> kernel/include/fs/fs.h 	renamed:    src/kernel/include/fs/msdospart.h -> kernel/include/fs/msdospart.h 	renamed:    src/kernel/include/memory.h -> kernel/include/memory.h 	renamed:    src/libc/include/string.h -> kernel/include/string.h 	renamed:    src/kernel/kernel.c -> kernel/kernel.c 	renamed:    linker.ld -> kernel/linker.ld 	renamed:    src/kernel/memory/mmu.c -> kernel/memory/mmu.c 	renamed:    src/kernel/memory/paging.c -> kernel/memory/paging.c 	renamed:    src/kernel/display.c -> kernel/utils/display.c 	renamed:    src/libc/src/dsa/hashmap.c -> kernel/utils/dsa/hashmap.c 	renamed:    src/libc/src/dsa/list.c -> kernel/utils/dsa/list.c 	renamed:    src/libc/src/dsa/tree.c -> kernel/utils/dsa/tree.c 	renamed:    src/libc/src/string.c -> kernel/utils/string.c 	deleted:    src/boot/multiboot2.asm
	new file:   kernel/fs/ext2.c
	new file:   kernel/include/fs/ext2.h

+++ kernel/drivers/ata.c
@@ -1,6 +1,6 @@
 #include <drivers/ata.h>
 
-uint8_t *ide_buf = 0;
+uint16_t *ide_buf = 0;
 
 static void ata_primary_irq(regs_t *r) {
     irq_ack(ATA_PRIMARY_IRQ);
